pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "brain-helper"
version = "0.1.0"
description = "A helper for World Quant Brain"
requires-python = ">=3.12"
license = { file = "LICENSE" }
dependencies = [
    "pandas",
    "peewee",
    "psycopg2-binary",
    "tqdm",
    "pyyaml",
    "requests",
    "jinja2"
]

[project.urls]
Repository = "https://github.com/Lazybones3/brain-helper.git"

[project.scripts]
init_tables = "brain.database:init_tables"
